Seguence of button enable/disable for AnesthOpen, SurgOpen, SurgClose, AnesthClose, Anesthesia Score now functions correctly, particularly if anesthesia times have already been saved to db; some minor control alignment on both forms; error checking for fields dose and qtywasted on FormAnesthMedDelDose
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@2199 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDental/Forms/FormAnesthMedDelDose.cs
@@ -8,6 +8,7 @@ using System.Text;
 using System.Windows.Forms;
 using OpenDentBusiness;
 using OpenDental.UI;
+using System.Text.RegularExpressions;
 
 namespace OpenDental
 {
@@ -44,11 +45,28 @@ namespace OpenDental
 		}
 
 		private void butClose_Click(object sender, EventArgs e){
+			//Regular Expressions to validate the format of the entered value in the textAnesthDose and textQtyWasted
+			Regex regexDose = new Regex("^\\d{1,4}(\\.\\d{1,2})?$");
+			Regex regexQtyWasted = new Regex("^\\d{1,4}(\\.\\d{1,2})?$");
+			if (!(regexDose.IsMatch(textDose.Text)) && textDose.Text != "")
+			{
+				MessageBox.Show("Dose should be xxxx.xx format");
+				textQtyWasted.Focus();
+			}
+			
+			else if (!(regexQtyWasted.IsMatch(textQtyWasted.Text)) && textQtyWasted.Text != "")
+			{
+				MessageBox.Show("Amount wasted should be xxxx.xx format");
+				textQtyWasted.Focus();
+			}
+			else
+			{
+				double dose = Convert.ToDouble(textDose.Text);
+				double amtWasted = Convert.ToDouble(textQtyWasted.Text);
+				AMedications.UpdateAMedDose(textAnesthMedName.Text, Convert.ToDouble(textDose.Text), Convert.ToDouble(amtWasted), textDoseTimeStamp.Text, anestheticRecordNum);
+				Close();
+			}
 
-			double dose = Convert.ToDouble(textDose.Text);
-			double amtWasted = Convert.ToDouble(textQtyWasted.Text);
-			AMedications.UpdateAMedDose(textAnesthMedName.Text, Convert.ToDouble(textDose.Text), Convert.ToDouble(amtWasted), textDoseTimeStamp.Text, anestheticRecordNum);
-			Close();
 		}
 
 		private void textDose_TextChanged(object sender, EventArgs e){
@@ -56,10 +74,26 @@ namespace OpenDental
 		}
 
 		private void butDelAnesthMeds_Click(object sender, EventArgs e){
+			//Regular Expressions to validate the format of the entered value in the textAnesthDose and textQtyWasted
+			Regex regexDose = new Regex("^\\d{1,4}(\\.\\d{1,2})?$");
+			Regex regexQtyWasted = new Regex("^\\d{1,4}(\\.\\d{1,2})?$");
+			if (!(regexDose.IsMatch(textDose.Text)) && textDose.Text != "")
+			{
+				MessageBox.Show("Dose should be xxxx.xx format");
+				textQtyWasted.Focus();
+			}
 
-			AMedications.DeleteAMedDose(textAnesthMedName.Text, Convert.ToDouble(textDose.Text), Convert.ToDouble(textQtyWasted.Text), textDoseTimeStamp.Text, anestheticRecordNum);
-			
-			DialogResult = DialogResult.OK;
+			else if (!(regexQtyWasted.IsMatch(textQtyWasted.Text)) && textQtyWasted.Text != "")
+			{
+				MessageBox.Show("Amount wasted should be xxxx.xx format");
+				textQtyWasted.Focus();
+			}
+			else
+			{
+				AMedications.DeleteAMedDose(textAnesthMedName.Text, Convert.ToDouble(textDose.Text), Convert.ToDouble(textQtyWasted.Text), textDoseTimeStamp.Text, anestheticRecordNum);
+
+				DialogResult = DialogResult.OK;
+			}
 		}
 
 		private void groupBoxAnesthMedDelete_Enter(object sender, EventArgs e){
